requirements-dev.txt
-i https://pypi.tuna.tsinghua.edu.cn/simple
pre-commit==1.20.0
flake8==3.7.9
isort==4.3.21
black==20.8b1
pytest==6.2.2
mypy==0.812
python-dotenv==0.15.0
flask-debugtoolbar==0.11.0
pytest-cov==2.11.1